+++ .prettierrc
@@ -1,5 +1,5 @@
 {
     "singleQuote": true,
-    "printWidth": 100,
-    "tabWidth": 2
+    "printWidth": 160,
+    "tabWidth": 4
 }
