+++ src/components/MessageBar/index.tsx
@@ -10,79 +10,74 @@ import { RootState, MessageState } from '../../types/store';
 import { getMainNode } from '../../utils/settings';
 
 const SnackbarWrapper = styled(Snackbar)`
-  &&& {
-    &.MuiSnackbar-root {
-      min-width: 500px;
-      max-width: 1000px;
-      padding: 0;
+    &&& {
+        &.MuiSnackbar-root {
+            min-width: 500px;
+            max-width: 1000px;
+            padding: 0;
+        }
+        .MuiSnackbarContent-root {
+            padding: 0;
+            width: 100%;
+        }
+        .MuiSnackbarContent-message {
+            padding: 0;
+            width: 100%;
+        }
     }
-    .MuiSnackbarContent-root {
-      padding: 0;
-      width: 100%;
-    }
-    .MuiSnackbarContent-message {
-      padding: 0;
-      width: 100%;
-    }
-  }
 `;
 
 function MessageBar() {
-  const settings = useSelector<RootState, any>((state: RootState) => state.settings);
-  const { text, hash, isError, localeParam } = useSelector<RootState, MessageState>(
-    state => state.message,
-    shallowEqual
-  );
-  const dispatch = useDispatch();
-  const [open, setOpen] = useState(false);
+    const settings = useSelector<RootState, any>((state: RootState) => state.settings);
+    const { text, hash, isError, localeParam } = useSelector<RootState, MessageState>(state => state.message, shallowEqual);
+    const dispatch = useDispatch();
+    const [open, setOpen] = useState(false);
 
-  function openLink() {
-    const { selectedNode, nodesList } = settings;
-    const currentNode = getMainNode(nodesList, selectedNode);
+    function openLink() {
+        const { selectedNode, nodesList } = settings;
+        const currentNode = getMainNode(nodesList, selectedNode);
 
-    openLinkToBlockExplorer(hash, currentNode.network);
-    onClose();
-  }
-
-  function formatHash() {
-    if (hash.length > 10) {
-      return `${hash.slice(0, 4)}...${hash.slice(hash.length - 4, hash.length)}`;
+        openLinkToBlockExplorer(hash, currentNode.network);
+        onClose();
     }
-    return hash;
-  }
 
-  function onClose() {
-    setOpen(false);
-    setTimeout(() => {
-      dispatch(clearMessageAction());
-    }, 200);
-  }
-
-  useEffect(() => {
-    if (text) {
-      setOpen(true);
+    function formatHash() {
+        if (hash.length > 10) {
+            return `${hash.slice(0, 4)}...${hash.slice(hash.length - 4, hash.length)}`;
+        }
+        return hash;
     }
-  }, [text]);
 
-  console.log(`hash in popup '${hash}', '${text}'`);
+    function onClose() {
+        setOpen(false);
+        setTimeout(() => {
+            dispatch(clearMessageAction());
+        }, 200);
+    }
 
-  return (
-    <SnackbarWrapper
-      anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-      open={open}
-      onClose={() => onClose()}
-      message={
-        <MessageContent
-          content={text}
-          hash={formatHash()}
-          openLink={() => openLink()}
-          onClose={() => onClose()}
-          isError={isError}
-          localeParam={localeParam}
+    useEffect(() => {
+        if (text) {
+            setOpen(true);
+        }
+    }, [text]);
+
+    return (
+        <SnackbarWrapper
+            anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
+            open={open}
+            onClose={() => onClose()}
+            message={
+                <MessageContent
+                    content={text}
+                    hash={formatHash()}
+                    openLink={() => openLink()}
+                    onClose={() => onClose()}
+                    isError={isError}
+                    localeParam={localeParam}
+                />
+            }
         />
-      }
-    />
-  );
+    );
 }
 
 export default MessageBar;
