+++ tslint.json
@@ -13,7 +13,7 @@
     },
     "jsRules": {},
     "rules": {
-        "max-line-length": [true, 120],
+        "max-line-length": [true, 160],
         "ordered-imports": false,
         "object-literal-sort-keys": false,
         "trailing-comma": false,
